refactor(fx/tab): put hard coded code to resource file
