+++ tabs/src/components/code/tab/InitializeDefinitionCode.txt
@@ -1,0 +1,5 @@
+// @microsoft.teams-js
+// import
+import * as microsoftTeams from "@microsoft/teams-js";
+// function
+function initialize(callback?: () => void, validMessageOrigins?: string[])
