+++ tabs/src/components/code/tab/deepLink/DefinitionCode.txt
@@ -1,0 +1,2 @@
+// @microsoft.teams-js
+function executeDeepLink(deepLink: string, onComplete?: (status: boolean, reason?: string) => void)
