+++ tabs/src/components/code/tab/staticTab/JsxLocationCode.txt
@@ -1,0 +1,1 @@
+./src/<YourPath>/PersonalTab.jsx
