+++ tabs/src/components/code/tab/staticTab/ManifestLocationCode.txt
@@ -1,0 +1,1 @@
+./src/manifest/manifest.json/
