+++ tabs/src/components/code/tab/staticTab/PersonalTabRouteCode.txt
@@ -1,0 +1,2 @@
+import PersonalTab from "./<YourPath>/PersonalTab";
+<Route exact path="/personaltab" component={ PersonalTab } />'
