+++ tabs/src/components/teamsfx/Tab.jsx
@@ -2,17 +2,16 @@ import React from "react";
 import { Button, Header, Text } from '@fluentui/react-northstar'
 import Deeplink from './tab/Deeplink.jsx'
 import StaticTab from './tab/StaticTab.jsx'
+import InitializeDefinition from '!!raw-loader!../code/tab/InitializeDefinitionCode.txt'
 import { TopJumper, ScrollToAnchor } from "../util/ScrollUtil";
 import Code from "../util/CodeUtil"
 
 function InitializeCode() {
-  const initializeDefinition = '// @microsoft.teams-js\n// import * as microsoftTeams from "@microsoft/teams-js";\nfunction initialize(callback?: () => void, validMessageOrigins?: string[])'
-
   return (
     <div>
       <Header as="h4" content="Initializes the library" />
       <Text content='This must be called before any other SDK calls but after the frame is loaded successfully.' />
-      <Code code={ initializeDefinition } />
+      <Code code={ InitializeDefinition } />
     </div>
   )
 }
